Rename classes
Renamed controllers with the "Controller" suffix.
Created the Repositories namespace with the corresponding classes.
