+++ public/index.php
@@ -18,12 +18,14 @@ $dotenv = Dotenv\Dotenv::createImmutable(ROOT, "/.env");
 $dotenv->load();
 
 use App\Core\Router;
-use App\Controllers\Homepage;
-use App\Controllers\Error;
+use App\Controllers\HomepageController;
+use App\Controllers\ContactFormController;
+use App\Controllers\ErrorController;
 use App\Core\Exceptions\Client\ClientException;
 use App\Core\Exceptions\Server\ServerException;
 
 $routes = [
+    "/" => fn () => (new HomepageController())->show(),
     "/" => fn () => (new Homepage())->show(),
 ];
 
@@ -31,11 +33,11 @@ try {
     $router = new Router($routes);
     $router->match();
 } catch (ClientException $e) {
-    (new Error($e))->show();
+    (new ErrorController($e))->show();
 } catch (ServerException $e) {
     error_logger($e);
-    (new Error($e))->show();
+    (new ErrorController($e))->show();
 } catch (\Throwable $th) {
     error_logger($th);
-    (new Error(new ServerException))->show();
+    (new ErrorController(new ServerException))->show();
 }
